Rename junit rules package
Authored-by: Jianxia Chen <jchen21@apache.org>
