+++ geode-connectors/src/acceptanceTest/java/org/apache/geode/connectors/jdbc/PostgresJdbcAsyncWriterIntegrationTest.java
@@ -20,8 +20,8 @@ import java.sql.SQLException;
 
 import org.junit.ClassRule;
 
-import org.apache.geode.connectors.jdbc.rules.DatabaseConnectionRule;
-import org.apache.geode.connectors.jdbc.rules.PostgresConnectionRule;
+import org.apache.geode.connectors.jdbc.test.junit.rules.DatabaseConnectionRule;
+import org.apache.geode.connectors.jdbc.test.junit.rules.PostgresConnectionRule;
 
 public class PostgresJdbcAsyncWriterIntegrationTest extends JdbcAsyncWriterIntegrationTest {
