+++ geode-connectors/src/acceptanceTest/java/org/apache/geode/connectors/jdbc/PostgresJdbcDistributedTest.java
@@ -23,8 +23,8 @@ import java.sql.Types;
 
 import org.junit.ClassRule;
 
-import org.apache.geode.connectors.jdbc.rules.PostgresConnectionRule;
-import org.apache.geode.connectors.jdbc.rules.SqlDatabaseConnectionRule;
+import org.apache.geode.connectors.jdbc.test.junit.rules.PostgresConnectionRule;
+import org.apache.geode.connectors.jdbc.test.junit.rules.SqlDatabaseConnectionRule;
 
 /**
  * End-to-end dunits for jdbc connector
