+++ geode-connectors/src/acceptanceTest/java/org/apache/geode/connectors/jdbc/PostgresJdbcLoaderIntegrationTest.java
@@ -29,8 +29,8 @@ import org.junit.Test;
 
 import org.apache.geode.cache.Region;
 import org.apache.geode.connectors.jdbc.internal.configuration.FieldMapping;
-import org.apache.geode.connectors.jdbc.rules.DatabaseConnectionRule;
-import org.apache.geode.connectors.jdbc.rules.PostgresConnectionRule;
+import org.apache.geode.connectors.jdbc.test.junit.rules.DatabaseConnectionRule;
+import org.apache.geode.connectors.jdbc.test.junit.rules.PostgresConnectionRule;
 import org.apache.geode.pdx.FieldType;
 import org.apache.geode.pdx.PdxInstance;
