+++ geode-connectors/src/acceptanceTest/java/org/apache/geode/connectors/jdbc/internal/PostgresTableMetaDataManagerIntegrationTest.java
@@ -27,8 +27,8 @@ import org.junit.ClassRule;
 import org.junit.Test;
 
 import org.apache.geode.connectors.jdbc.internal.configuration.RegionMapping;
-import org.apache.geode.connectors.jdbc.rules.DatabaseConnectionRule;
-import org.apache.geode.connectors.jdbc.rules.PostgresConnectionRule;
+import org.apache.geode.connectors.jdbc.test.junit.rules.DatabaseConnectionRule;
+import org.apache.geode.connectors.jdbc.test.junit.rules.PostgresConnectionRule;
 
 public class PostgresTableMetaDataManagerIntegrationTest
     extends TableMetaDataManagerIntegrationTest {
