+++ geode-connectors/src/acceptanceTest/java/org/apache/geode/connectors/jdbc/test/junit/rules/InMemoryDerbyConnectionRule.java
@@ -12,15 +12,31 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.apache.geode.connectors.jdbc.rules;
+package org.apache.geode.connectors.jdbc.test.junit.rules;
 
 import java.sql.Connection;
+import java.sql.DriverManager;
 import java.sql.SQLException;
 
-import org.junit.rules.TestRule;
+import org.junit.rules.ExternalResource;
 
-public interface DatabaseConnectionRule extends TestRule {
-  Connection getConnection() throws SQLException;
+public class InMemoryDerbyConnectionRule extends ExternalResource
+    implements DatabaseConnectionRule {
+  private static final String CONNECTION_URL = "jdbc:derby:memory:%s;create=true";
 
-  String getConnectionUrl();
+  private final String dbName;
+
+  public InMemoryDerbyConnectionRule(String dbName) {
+    this.dbName = dbName;
+  }
+
+  @Override
+  public Connection getConnection() throws SQLException {
+    return DriverManager.getConnection(getConnectionUrl());
+  }
+
+  @Override
+  public String getConnectionUrl() {
+    return String.format(CONNECTION_URL, dbName);
+  }
 }
